+++ src/components/InstancesList/InstancesList.js
@@ -5,6 +5,7 @@ import {
   get,
   set,
   flowRight,
+  size,
   isEmpty,
   noop,
 } from 'lodash';
@@ -40,7 +41,10 @@ import {
   marshalInstance,
   omitFromArray,
 } from '../../utils';
-import { INSTANCES_ID_REPORT_TIMEOUT } from '../../constants';
+import {
+  INSTANCES_ID_REPORT_TIMEOUT,
+  QUICK_EXPORT_LIMIT,
+} from '../../constants';
 import {
   InTransitItemReport,
   InstancesIdReport,
@@ -273,7 +277,9 @@ class InstancesView extends React.Component {
 
   getActionMenu = ({ onToggle }) => {
     const { parentResources } = this.props;
+    const selectedRowsCount = size(this.state.selectedRows);
     const isInstancesListEmpty = isEmpty(get(parentResources, ['records', 'records'], []));
+    const isQuickExportLimitExceeded = selectedRowsCount > QUICK_EXPORT_LIMIT;
 
     const buildOnClickHandler = onClickHandler => {
       return () => {
@@ -332,8 +338,19 @@ class InstancesView extends React.Component {
           icon: 'download',
           messageId: 'ui-inventory.exportInstancesInMARC',
           onClickHandler: buildOnClickHandler(noop),
-          isDisabled: true,
+          isDisabled: !selectedRowsCount || isQuickExportLimitExceeded,
         })}
+        {isQuickExportLimitExceeded && (
+          <span
+            className={css.feedbackError}
+            data-test-quick-marc-export-limit-exceeded
+          >
+            <FormattedMessage
+              id="ui-inventory.exportInstancesInMARCLimitExceeded"
+              values={{ count: QUICK_EXPORT_LIMIT }}
+            />
+          </span>
+        )}
         {this.getActionItem({
           id: 'dropdown-clickable-export-json',
           icon: 'download',
@@ -366,6 +383,23 @@ class InstancesView extends React.Component {
 
   handleResetAll = () => {
     this.setState({ selectedRows: {} });
+  }
+
+  renderPaneSub() {
+    const selectedRowsCount = size(this.state.selectedRows);
+
+    return selectedRowsCount
+      ? (
+        <FormattedMessage
+          id="ui-inventory.instances.rows.recordsSelected"
+          values={{ count: selectedRowsCount }}
+        />
+      )
+      : null;
+  }
+
+  formatCellStyles(defaultCellStyle) {
+    return `${defaultCellStyle} ${css.cellAlign}`;
   }
 
   render() {
@@ -453,6 +487,8 @@ class InstancesView extends React.Component {
               select: '30px',
               title: '40%',
             }}
+            getCellClass={this.formatCellStyles}
+            customPaneSub={this.renderPaneSub()}
             resultsFormatter={resultsFormatter}
             onCreate={this.onCreate}
             viewRecordPerms="ui-inventory.instance.view"