+++ test/bigtest/interactors/routes/instances-route.js
@@ -17,4 +17,5 @@ export default @interactor class InstancesRouteInteractor {
   rows = collection('#list-inventory [data-row-index]');
   headers = collection('#list-inventory [data-test-clickable-header]');
   selectRowCheckboxes = collection('[data-test-inventory-instances] [data-test-select-row]', CheckboxInteractor);
+  customPaneSub = scoped('[data-test-custom-pane-sub]');
 }